Selected option fixes for Safari and Firefox
In the case when the first option is a placeholder which has an empty value and data-color and the select option is dynamically populated from a database or restful service,  the selected option is not updated in Safari and Firefox. In Safari, the first and current option is selected. In Firefox, all options left deselected. To fix this, I set the selectedIndex to -1 for Safari and for Firefox I set the select's value to the colorDiv value.